rearrage files and split logic
