+++ .github/workflows/jira-snyk.yaml
@@ -33,4 +33,4 @@ jobs:
           pip install -r requirements.txt
 
       - name: execute py script for automation 
-        run: python scripts/jira-automation.py
+        run: python jira-automation.py
