+++ .gitignore
@@ -1,1 +1,1 @@
-__pycache__
+__pycache__/
